Fixed memory leaks

+++ Source/Func/IuCommonFunctions.cpp
@@ -184,12 +184,12 @@ CString GenerateFileName(const CString& templateStr, int index, const CPoint siz
     result.Replace(_T("%md5"), (LPCTSTR)md5);
     result.Replace(_T("%width%"), WinUtils::IntToStr(size.x));
     result.Replace(_T("%height%"), WinUtils::IntToStr(size.y));
-    year.Format(_T("%04d"), (int)1900 + timeinfo->tm_year);
-    month.Format(_T("%02d"), (int)timeinfo->tm_mon + 1);
-    day.Format(_T("%02d"), (int)timeinfo->tm_mday);
-    hours.Format(_T("%02d"), (int)timeinfo->tm_hour);
-    seconds.Format(_T("%02d"), (int)timeinfo->tm_sec);
-    minutes.Format(_T("%02d"), (int)timeinfo->tm_min);
+    year.Format(_T("%04d"), 1900 + timeinfo->tm_year);
+    month.Format(_T("%02d"), timeinfo->tm_mon + 1);
+    day.Format(_T("%02d"), timeinfo->tm_mday);
+    hours.Format(_T("%02d"), timeinfo->tm_hour);
+    seconds.Format(_T("%02d"), timeinfo->tm_sec);
+    minutes.Format(_T("%02d"), timeinfo->tm_min);
     result.Replace(_T("%y"), year);
     result.Replace(_T("%m"), month);
     result.Replace(_T("%d"), day);
